финальная версия задания 9.1
