+++ package.json
@@ -17,7 +17,6 @@
     "eslint-config-htmlacademy": "2.0.0",
     "file-loader": "6.2.0",
     "mini-css-extract-plugin": "1.3.9",
-    "style-loader": "2.0.0",
     "webpack": "5.24.4",
     "webpack-cli": "4.5.0"
   },
