+++ source/js/map.js
@@ -1,6 +1,6 @@
 import L from 'leaflet';
 import 'leaflet/dist/leaflet.css';
-import {setAddress} from './form.js';
+import {setAddress} from './form/form.js'
 import {createCard} from './create-card.js';
 
 const MAX_ADS_COUNT = 10;
